requirements.txt

keyboard~=0.13.5
pystray~=0.19.4
Pillow==9.4.0
PyAutoGUI~=0.9.53